requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.4
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
asyncpg==0.30.0
    # via cv-analysis
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.13.4
    # via gdown
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   zstandard
charset-normalizer==3.4.2
    # via
    #   pdfminer-six
    #   requests
click==8.2.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
cryptography==43.0.3
    # via pdfminer-six
dataclasses-json==0.6.7
    # via langchain-community
fastapi==0.115.12
    # via cv-analysis
filelock==3.18.0
    # via gdown
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
gdown==5.2.0
    # via cv-analysis
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.24.2
    # via google-ai-generativelanguage
google-auth==2.40.2
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-genai
google-genai==1.18.0
    # via cv-analysis
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.2 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
    #   qdrant-client
grpcio-status==1.71.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   langsmith
    #   qdrant-client
httpx-sse==0.4.0
    # via langchain-community
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via
    #   cv-analysis
    #   langchain-community
langchain-community==0.3.24
    # via cv-analysis
langchain-core==0.3.62
    # via
    #   cv-analysis
    #   langchain
    #   langchain-community
    #   langchain-google-genai
    #   langchain-text-splitters
langchain-google-genai==2.1.5
    # via cv-analysis
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.43
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via
    #   langchain-community
    #   qdrant-client
orjson==3.10.18 ; platform_python_implementation != 'PyPy'
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pdfminer-six==20250327
    # via pdfplumber
pdfplumber==0.11.6
    # via cv-analysis
pillow==11.2.1
    # via pdfplumber
portalocker==2.10.1
    # via qdrant-client
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   qdrant-client
psycopg2-binary==2.9.10
    # via cv-analysis
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
    #   pydantic-settings
    #   qdrant-client
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via langchain-community
pypdfium2==4.30.1
    # via pdfplumber
pysocks==1.7.1
    # via requests
python-dotenv==1.1.0
    # via
    #   cv-analysis
    #   pydantic-settings
python-multipart==0.0.20
    # via cv-analysis
pywin32==310 ; sys_platform == 'win32'
    # via portalocker
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
qdrant-client==1.14.2
    # via cv-analysis
requests==2.32.3
    # via
    #   cv-analysis
    #   gdown
    #   google-api-core
    #   google-genai
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tqdm==4.67.1
    # via gdown
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   google-genai
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   qdrant-client
    #   requests
uvicorn==0.34.3
    # via cv-analysis
websockets==15.0.1
    # via google-genai
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
